requirements.txt
anyio==4.3.0
    # via watchfiles
blinker==1.7.0
    # via pelican
docutils==0.21.1
    # via pelican
feedgenerator==2.1.0
    # via pelican
ghp-import==2.1.0
    # via -r requirements.in
idna==3.7
    # via anyio
jinja2==3.1.3
    # via pelican
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
ordered-set==4.1.0
    # via pelican
pelican==4.9.1
    # via -r requirements.in
pillow==10.3.0
    # via -r requirements.in
pygments==2.17.2
    # via
    #   pelican
    #   rich
python-dateutil==2.9.0.post0
    # via
    #   ghp-import
    #   pelican
pytz==2024.1
    # via feedgenerator
rich==13.7.1
    # via pelican
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
unidecode==1.3.8
    # via pelican
watchfiles==0.21.0
    # via pelican
